refactor: Update ZenKeyboardShortcuts to include split view and workspace actions
This commit updates the ZenKeyboardShortcuts module to include new actions for split view and workspace management. It adds shortcuts for toggling between grid, vertical, and horizontal split views, as well as closing split views. Additionally, it introduces a shortcut for changing workspaces.

The changes also include the addition of default shortcuts for the new actions and the initialization of shortcuts only on the browser page.

Refactor other modules:
- Remove unnecessary console.log statement in ZenThemeBuilder
- Disable Firefox shortcuts in ZenKeyboardShortcuts
- Remove redundant code in ZenViewSplitter

Refactor ZenProfileDialogUI:
- Improve profiles list and profile creation

Refactor ZenWorkspaces.mjs:
- Remove unnecessary code

Refactor ZenViewSplitter.mjs:
- Activate docShell for tab
- Remove redundant code